# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WorkForcePro API",
    "version": "1.0.0",
    "description": "RESTful API for managing employees, trainings, certifications, and departments in a workforce management system."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/employees/{id}/trainings": {
      "get": {
        "tags": ["Employees", "Trainings"],
        "summary": "Get trainings for an employee",
        "operationId": "getEmployeeTrainings",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeId" }
        ],
        "responses": {
          "200": {
            "description": "Employee training assignments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EmployeeTraining" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/trainings": {
      "get": {
        "tags": ["Trainings"],
        "summary": "List all trainings",
        "operationId": "getTrainings",
        "responses": {
          "200": {
            "description": "List of training modules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Training" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/employee/{employee_id}": {
      "get": {
        "tags": ["Trainings"],
        "summary": "Get trainings for a specific employee",
        "operationId": "getTrainingsByEmployee",
        "parameters": [
          { "$ref": "#/components/parameters/EmployeeNumericId" }
        ],
        "responses": {
          "200": {
            "description": "Employee training assignments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/EmployeeTraining" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/assign": {
      "post": {
        "tags": ["Trainings"],
        "summary": "Assign training to an employee (supervisor only)",
        "operationId": "assignTraining",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssignTrainingRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Training assigned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmployeeTraining" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": {
            "description": "Only supervisors can assign training",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/complete": {
      "post": {
        "tags": ["Trainings"],
        "summary": "Mark training as completed",
        "operationId": "completeTraining",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompleteTrainingRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Training marked as completed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmployeeTraining" }
              }
            }
          },
          "201": {
            "description": "Training completed without prior assignment",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/EmployeeTraining" },
                    {
                      "type": "object",
                      "properties": {
                        "warning": {
                          "type": "string",
                          "example": "Training was not assigned but was completed anyway"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/approve": {
      "post": {
        "tags": ["Trainings"],
        "summary": "Approve completed training (supervisor only)",
        "operationId": "approveTraining",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TrainingActionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Training approved",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmployeeTraining" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": {
            "description": "Only supervisors can approve training",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/reject": {
      "post": {
        "tags": ["Trainings"],
        "summary": "Reject training (supervisor only)",
        "operationId": "rejectTraining",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TrainingActionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Training rejected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Training rejected" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "403": {
            "description": "Only supervisors can reject training",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/{id}": {
      "get": {
        "tags": ["Trainings"],
        "summary": "Get a single training module",
        "operationId": "getTraining",
        "parameters": [
          { "$ref": "#/components/parameters/TrainingId" }
        ],
        "responses": {
          "200": {
            "description": "Training details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Training" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/trainings/{id}/assignments": {
      "get": {
        "tags": ["Trainings"],
        "summary": "Get all employees assigned to a training",
        "operationId": "getTrainingAssignments",
        "parameters": [
          { "$ref": "#/components/parameters/TrainingId" }
        ],
        "responses": {
          "200": {
            "description": "List of assignments for the training",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/TrainingAssignment" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "EmployeeId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Employee database ID"
      },
      "EmployeeNumericId": {
        "name": "employee_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Employee database ID"
      },
      "TrainingId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" },
        "description": "Training database ID"
      },
...
    },
    "schemas": {
...
      "Training": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "training_name": { "type": "string", "example": "Workplace Safety" },
          "description": { "type": "string", "nullable": true, "example": "Annual workplace safety training" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "EmployeeTraining": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "employee_id": { "type": "integer", "example": 2 },
          "training_id": { "type": "integer", "example": 1 },
          "training_name": { "type": "string", "example": "Workplace Safety" },
          "due_date": { "type": "string", "format": "date", "nullable": true, "example": "2026-07-25" },
          "status": {
            "type": "string",
            "enum": ["assigned", "completed", "approved", "rejected"],
            "example": "assigned"
          },
          "completion_date": { "type": "string", "format": "date", "nullable": true, "example": "2026-07-20" },
          "notes": { "type": "string", "nullable": true, "example": "Completed all modules" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "AssignTrainingRequest": {
        "type": "object",
        "required": ["employee_id", "training_id", "supervisor_id"],
        "properties": {
          "employee_id": { "type": "integer", "example": 2 },
          "training_id": { "type": "integer", "example": 1 },
          "due_date": { "type": "string", "format": "date", "example": "2026-07-25" },
          "supervisor_id": { "type": "integer", "example": 3 }
        }
      },
      "CompleteTrainingRequest": {
        "type": "object",
        "required": ["employee_id", "training_id"],
        "properties": {
          "employee_id": { "type": "integer", "example": 2 },
          "training_id": { "type": "integer", "example": 1 },
          "completion_date": { "type": "string", "format": "date", "example": "2026-06-25" },
          "notes": { "type": "string", "example": "Completed all modules successfully" }
        }
      },
      "TrainingActionRequest": {
        "type": "object",
        "required": ["employee_id", "training_id", "supervisor_id"],
        "properties": {
          "employee_id": { "type": "integer", "example": 2 },
          "training_id": { "type": "integer", "example": 1 },
          "supervisor_id": { "type": "integer", "example": 3 }
        }
      },
      "TrainingAssignment": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 5 },
          "employee_id": { "type": "integer", "example": 2 },
          "employee_name": { "type": "string", "example": "John Doe" },
          "training_id": { "type": "integer", "example": 1 },
          "due_date": { "type": "string", "format": "date", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["assigned", "completed", "approved", "rejected"],
            "example": "assigned"
          },
          "completion_date": { "type": "string", "format": "date", "nullable": true },
          "notes": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Employee not found" }
        }
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  }
}